+++ packages/ramps-controller/CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixed
+### Changed
 
-- Fix `getWidgetUrl()` to extract widget URL from `quote.quote.buyWidget.url` instead of top-level `quote.url` ([#7881](https://github.com/MetaMask/core/pull/7881))
+- **BREAKING:** Make `getWidgetUrl()` async to fetch the actual provider widget URL from the `buyURL` endpoint ([#7881](https://github.com/MetaMask/core/pull/7881))
 
 ## [7.1.0]